feat: resource references can be declared in any order (backward or forward).
